Update AssetRegistryDependencyDataOffset when any table size changes
There is an offset stored in AssetRegistryData, the
AssetRegistryDependencyDataOffset. It is not correctly updated if any
table precedes it has expanded or shrunk. See ReadPackageDataMain and
WritePackageData in Unreal Engine code for more detail about this
offset. This patch may not work for all cases, since it does not
replicate the computation of the offset. It is based on the assumption
that the difference between AssetRegistryDataOffset and
AssetRegistryDependencyDataOffset usually does not change.